add fast wirter support
