+++ src/main/java/com/taosdata/jdbc/AbstractResultSet.java
@@ -1,5 +1,7 @@
 package com.taosdata.jdbc;
 
+import com.taosdata.jdbc.utils.Utils;
+
 import java.io.InputStream;
 import java.io.Reader;
 import java.math.BigDecimal;
@@ -13,9 +15,6 @@ public abstract class AbstractResultSet extends WrapperImpl implements ResultSet
     private int fetchSize;
     protected boolean wasNull;
     protected int timestampPrecision;
-
-    private static final ForkJoinPool forkJoinPool = new ForkJoinPool();
-
     public void setTimestampPrecision(int timestampPrecision) {
         this.timestampPrecision = timestampPrecision;
     }
@@ -31,7 +30,7 @@ public abstract class AbstractResultSet extends WrapperImpl implements ResultSet
 
 
     protected ForkJoinPool getForkJoinPool(){
-        return forkJoinPool;
+        return Utils.getForkJoinPool();
     }
 
     @Override
