+++ src/main/java/com/taosdata/jdbc/TSDBDriver.java
@@ -121,6 +121,14 @@ public class TSDBDriver extends AbstractDriver {
     public static final String PROPERTY_KEY_DISABLE_SSL_CERT_VALIDATION = "disableSSLCertValidation";
     public static final String PROPERTY_KEY_APP_IP = "app_ip";
     public static final String PROPERTY_KEY_APP_NAME = "app_name";
+    public static final String PROPERTY_KEY_LINGER_MS = "lingerMS";
+    public static final String PROPERTY_KEY_COPY_DATA = "copyData";
+    public static final String PROPERTY_KEY_BATCH_SIZE_BY_ROW = "batchSizeByRow";
+    public static final String PROPERTY_KEY_CACHE_SIZE_BY_ROW = "cacheSizeByRow";
+    public static final String PROPERTY_KEY_BACKEND_WRITE_THREAD_NUM = "backendWriteThreadNum";
+    public static final String PROPERTY_KEY_STRICT_CHECK = "strictCheck";
+    public static final String PROPERTY_KEY_RETRY_TIMES = "retryTimes";
+    public static final String PROPERTY_KEY_ASYNC_WRITE = "asyncWrite";
 
 
     /**
