+++ src/main/java/com/taosdata/jdbc/common/SerializeBlock.java
@@ -1,13 +1,11 @@
 package com.taosdata.jdbc.common;
 
-import com.taosdata.jdbc.TSDBConstants;
 import com.taosdata.jdbc.TSDBError;
 import com.taosdata.jdbc.TSDBErrorNumbers;
 import com.taosdata.jdbc.utils.DateTimeUtils;
 import com.taosdata.jdbc.utils.StringUtils;
 
 import java.io.ByteArrayOutputStream;
-import java.io.IOException;
 import java.math.BigInteger;
 import java.sql.SQLException;
 import java.time.Instant;
@@ -72,7 +70,7 @@ public class SerializeBlock {
         System.arraycopy(data, 0, buf, offset, data.length);
     }
 
-    private static int serializeColumn(ColumnInfo columnInfo, byte[] buf, int offset, int precision) throws IOException, SQLException {
+    private static int serializeColumn(ColumnInfo columnInfo, byte[] buf, int offset, int precision) throws SQLException {
         Integer dataLen = DataLengthCfg.getDataLength(columnInfo.getType());
 
         // TotalLength
@@ -147,7 +145,7 @@ public class SerializeBlock {
         return offset;
     }
 
-    private static void SerializeNormalDataType(int dataType , byte[] buf, int offset, List<Object> objectList, int precision) throws IOException, SQLException {
+    private static void SerializeNormalDataType(int dataType , byte[] buf, int offset, List<Object> objectList, int precision) throws SQLException {
         switch (dataType) {
             case TSDB_DATA_TYPE_BOOL: {
                 SerializeInt(buf, offset, objectList.size());
@@ -475,7 +473,7 @@ public class SerializeBlock {
                                            int toBeBindTableNameIndex,
                                            int toBebindTagCount,
                                            int toBebindColCount,
-                                           int precision) throws IOException, SQLException {
+                                           int precision) throws SQLException {
 
         // cloc totol size
         int totalTableNameSize  = 0;
