+++ src/main/java/com/taosdata/jdbc/ws/WSConnection.java
@@ -4,6 +4,7 @@ import com.taosdata.jdbc.AbstractConnection;
 import com.taosdata.jdbc.TSDBDriver;
 import com.taosdata.jdbc.TSDBError;
 import com.taosdata.jdbc.TSDBErrorNumbers;
+import com.taosdata.jdbc.enums.FeildBindType;
 import com.taosdata.jdbc.enums.SchemalessProtocolType;
 import com.taosdata.jdbc.enums.SchemalessTimestampType;
 import com.taosdata.jdbc.rs.ConnectionParam;
@@ -14,6 +15,10 @@ import com.taosdata.jdbc.ws.entity.*;
 import com.taosdata.jdbc.ws.entity.CommonResp;
 import com.taosdata.jdbc.ws.schemaless.InsertReq;
 import com.taosdata.jdbc.ws.schemaless.SchemalessAction;
+import com.taosdata.jdbc.ws.stmt2.entity.Field;
+import com.taosdata.jdbc.ws.stmt2.entity.RequestFactory;
+import com.taosdata.jdbc.ws.stmt2.entity.Stmt2PrepareResp;
+import com.taosdata.jdbc.ws.stmt2.entity.Stmt2Resp;
 
 import java.sql.DatabaseMetaData;
 import java.sql.PreparedStatement;
@@ -63,13 +68,49 @@ public class WSConnection extends AbstractConnection {
             database = this.getClientInfo(TSDBDriver.PROPERTY_KEY_DBNAME);
 
         if (transport != null && !transport.isClosed()) {
-            return new TSWSPreparedStatement(transport,
-                    param,
-                    database,
-                    this,
-                    sql,
-                    idGenerator.getAndIncrement(),
-                    param.getZoneId());
+            long reqId = ReqId.getReqID();
+            Request request = com.taosdata.jdbc.ws.stmt2.entity.RequestFactory.generateInit(reqId, true, true);
+            Stmt2Resp resp = (Stmt2Resp) transport.send(request);
+            if (Code.SUCCESS.getCode() != resp.getCode()) {
+                throw new SQLException("(0x" + Integer.toHexString(resp.getCode()) + "):" + resp.getMessage());
+            }
+            long stmtId = resp.getStmtId();
+            Request prepare = RequestFactory.generatePrepare(stmtId, reqId, sql);
+            Stmt2PrepareResp prepareResp = (Stmt2PrepareResp) transport.send(prepare);
+            if (Code.SUCCESS.getCode() != prepareResp.getCode()) {
+                Request close = RequestFactory.generateClose(stmtId, reqId);
+                transport.sendWithoutResponse(close);
+                throw new SQLException("(0x" + Integer.toHexString(prepareResp.getCode()) + "):" + prepareResp.getMessage());
+            }
+
+            boolean isInsert = prepareResp.isInsert();
+            boolean isSuperTable = false;
+            if (isInsert){
+                for (Field field : prepareResp.getFields()){
+                    if (field.getBindType() == FeildBindType.TAOS_FIELD_TBNAME.getValue()){
+                        isSuperTable = true;
+                        break;
+                    }
+                }
+            }
+
+            if ((sql.startsWith("ASYNC_INSERT") || "STMT".equalsIgnoreCase(param.getAsyncWrite())) && isInsert && isSuperTable) {
+                return new WSFastWriterPreparedStatement(transport,
+                        param,
+                        database,
+                        this,
+                        sql,
+                        idGenerator.getAndIncrement(),
+                        prepareResp);
+            } else {
+                return new TSWSPreparedStatement(transport,
+                        param,
+                        database,
+                        this,
+                        sql,
+                        idGenerator.getAndIncrement(),
+                        prepareResp);
+            }
         } else {
             throw TSDBError.createSQLException(TSDBErrorNumbers.ERROR_CONNECTION_CLOSED);
         }
